Metric prefixes

These functions simplify scaling between different metric prefixes.

scale_metric(q, prefix = "")

best_metric(q)

base_metric(q)

Arguments

q

the quantity to scale

prefix

a metric prefix (p, n, µ, m, k, M, etc.)

Functions

scale_metric(): scale to a specific metrix prefix (from whatever the quantity is currently in)

best_metric(): convert to best metric prefix (i.e. one that gives at least 1 significant digit before the decimal), if the quantity has a vector of values, scales to the best metric prefix for the median of all values

base_metric(): convert to base metric prefix of microbialkitchen (i.e. to mol, L, etc.)

See also